GOD DAMN MOTHER FUCKER SHIT FUCK CUNT BITCH Fixes a random map crash that I wasted a bunch of time on a few months ago.
Icemap loading uses calloc (as does creating a blank map from lua, because
reasons) so that a bunch of fields are pre-set to null in the case that we
add more. VXL loading however still uses malloc, and since VXL maps don't
have MapEnts, we end up with a gibberish pointer.