Replace GLEW dependency with Glad and set the bar at 2.0. Fixes #229 and #228
This removes GLEW as a dependency and adds [Glad](http://glad.dav1d.de/)
instead. We are also bumping the OpenGL version to 2.0.

This means we are doing a few things differently now:
* We ignore the `vbo` and `gl_vbo` options, VBOs are "core" in 2.0.
* We show a message box on startup on version mismatch instead of a
simple `printf`.
* We remove `gl_max_texunits`. This was used incorrectly. If we want to
ever go beyond `VA_MAX_IMG` (8), look into
`GL_MAX_COMBINED_TEXTURE_IMAGE_UNITS`.

If more extensions are required simply generate a new loader and replace
the files. The extensions that Glad (3.3 compat) was generated with are
the following:
* GL_ARB_debug_output,
* GL_ARB_framebuffer_object
* GL_ARB_framebuffer_sRGB
* GL_ARB_occlusion_query
* GL_EXT_framebuffer_object
* GL_EXT_framebuffer_sRGB
* GL_KHR_debug